When deleting an assembly before re-compile, make sure its attributes allow deletion.
This is to see if this helps with the problem in http://opensimulator.org/mantis/view.php?id=7278 where some DLLs are not allowing this.
Since OpenSim created the file it should always be allowed to delete it.

+++ OpenSim/Region/ScriptEngine/Shared/CodeTools/Compiler.cs
@@ -469,7 +469,11 @@ namespace SecondLife
             scriptCompileCounter++;
             try
             {
-                File.Delete(assembly);
+                if (File.Exists(assembly))
+                {
+                    File.SetAttributes(assembly, FileAttributes.Normal);
+                    File.Delete(assembly);
+                }
             }
             catch (Exception e) // NOTLEGIT - Should be just FileIOException
             {
